pyproject.toml
[project]
name = "arxiv_daily"
version = "0.1.0"
description = "ArXiv daily paper tracking and analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "litellm>=1.68.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.ruff]
line-length = 88
target-version = "py310"

